feat: merge Agent and MCP log panels into unified Activity Logs panel
Replace the two separate bottom tool windows (DevoxxGenieAgentLogs and
DevoxxGenieMCPLogs) with a single 'DevoxxGenie Logs' panel that combines
both log streams.

- Add AgentMcpLogPanel with LogSource (MCP/AGENT) tagging per entry
- Add filter dropdown: Show All / Show MCP Only / Show Agents Only
- Show [MCP]/[AGT] source badges in 'Show All' mode
- Preserve per-source colour coding (agent type colours + MCP direction colours)
- Double-click opens Agent entries as .txt, MCP entries as pretty-printed .json
- Remove misleading start/stop toggle buttons that disabled MCP/Agent functionality
- Register single DevoxxGenieActivityLogs toolWindow in plugin.xml with spy icon
- Add spy.svg / spy_dark.svg icons for the unified tool window
- Delete AgentLogPanel, MCPLogPanel, AgentLogToolWindowFactory, MCPLogToolWindowFactory
